Refactored models to domain layer, updated imports
