+++ app/main.py
@@ -2,7 +2,7 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI, HTTPException, Response
 from fastapi.middleware.cors import CORSMiddleware
 from app.constants.schemas import get_thought_schema
-from app.models.request import MessageRequest
+from app.domain.models import MessageRequest
 from app.services.thinking import periodic_thinking
 from app.services.openai_service import get_structured_response
 from app.services.data_service import get_all_agents, grab_self, init_db, db_client
